Annotate request with permission info in checkDeadline, display it
In the helper function for checking deadline satisfiability, if the user
is an administrator, annotate the request with information about which
permission or deadline was checked and what roles would have had
permission. Then display this information in the fruitsalad default
template.

To accommodate all this, also refactor deadline code to:

- add a new method for listing roles that have a particular permission,
  pulling out shared logic with user_has_perm
- eliminate duplication
- remove the unused meets_all_deadlines method.

Fixes #2324.