Add flag to prioritize students without classes
This might be kind of unfair since all of their requests are now
"stronger" than even the first request from a student with at least one
class. However, it's calculated dynamically so the advantage disappears
once they get their first class, and they don't get any extra priority
when we try to kick students, mostly just because that part of the code
is hard to reason about.